Support multiple structs
In the old implementation, enum for actor and action of each struct will be created
automatically from actor/action string with repr number value always start from 0.
This can cause the id conflict if the game use multiple structs to store spritesheet handles.

The new implementation uses xxhash to turn actor/action string into u64 number. So we can
have corresponding id across View2dCollection structs.

Use function 'get_act_id' to get number id from actor/action string.

+++ src/system.rs
@@ -5,7 +5,7 @@ use bevy::prelude::{Commands, Entity, MessageReader, MessageWriter, Query, Res, 
 
 use crate::component::*;
 
-const ACTION_ANY: u16 = 0;
+const ACTION_ANY: u64 = 0;
 const ACTOR_ANY: u64 = 0;
 
 /// Check on `ViewChanged` event and change to corresponding spritesheet.
